+++ app/libs/adblock/registerFiltering.js
@@ -1,49 +1,1 @@
-// https://github.com/minbrowser/min/blob/0a07a99415d8670e49c03a6e12d76a427784964e/main/filtering.js
-const { session } = require('electron');
-const fs = require('fs');
-const path = require('path');
-
-const extractDomain = (url) => {
-  const matches = url.match(/^https?:\/\/([^/?#]+)(?:[/?#]|$)/i);
-  const domain = matches && matches[1];
-  return domain;
-};
-
-const registerFiltering = (partitionId) => {
-  const { AdBlockClient, FilterOptions } = require('ad-block');
-
-  const filePath = path.join(__dirname, 'easylist+easyprivacy-noelementhiding.txt');
-  fs.readFile(filePath, 'utf8', (err, data) => {
-    if (err) return;
-
-    const client = new AdBlockClient();
-    client.parse(data);
-
-    session.fromPartition(`persist:${partitionId}`).webRequest
-      .onBeforeRequest((details, callback) => {
-        // main request
-        if (!(details.url.startsWith('http://') || details.url.startsWith('https://')) || details.resourceType === 'mainFrame') {
-          callback({
-            cancel: false,
-            requestHeaders: details.requestHeaders,
-          });
-          return;
-        }
-
-        if (client.matches(details.url, FilterOptions.script, extractDomain(details.referrer))) {
-          callback({
-            cancel: true,
-            requestHeaders: details.requestHeaders,
-          });
-          return;
-        }
-
-        callback({
-          cancel: false,
-          requestHeaders: details.requestHeaders,
-        });
-      });
-  });
-};
-
-module.exports = registerFiltering;
+const{session}=require('electron'),fs=require('fs'),path=require('path'),extractDomain=(a)=>{const b=a.match(/^https?:\/\/([^/?#]+)(?:[/?#]|$)/i),c=b&&b[1];return c},registerFiltering=(a)=>{const{AdBlockClient:b,FilterOptions:c}=require('ad-block'),d=path.join(__dirname,'easylist+easyprivacy-noelementhiding.txt');fs.readFile(d,'utf8',(e,f)=>{if(e)return;const g=new b;g.parse(f),session.fromPartition(`persist:${a}`).webRequest.onBeforeRequest((h,i)=>{return(h.url.startsWith('http://')||h.url.startsWith('https://'))&&'mainFrame'!==h.resourceType?g.matches(h.url,c.script,extractDomain(h.referrer))?void i({cancel:!0,requestHeaders:h.requestHeaders}):void i({cancel:!1,requestHeaders:h.requestHeaders}):void i({cancel:!1,requestHeaders:h.requestHeaders})})})};module.exports=registerFiltering;
