+++ package.json
@@ -6,11 +6,12 @@
   "scripts": {
     "postinstall": "install-app-deps",
     "start": "yarn run start.store",
-    "start.store": "electron ./app --development=true",
-    "start.app": "electron ./app --development=true --id=test --url=http://ads-blocker.com/testing/ --name=Shaka",
-    "start.store.production": "electron ./app",
-    "dev": "cross-env NODE_ENV=development webpack -d --watch",
-    "compile": "cross-env NODE_ENV=production webpack -p",
+    "start.store.dev": "electron ./app --development=true",
+    "start.app.dev": "electron ./app --development=true --id=test --url=http://ads-blocker.com/testing/ --name=Shaka",
+    "start.store": "electron ./app",
+    "start.app": "electron ./app --id=test --url=http://ads-blocker.com/testing/ --name=Shaka",
+    "dev": "babel src/main -d app --copy-files && cross-env NODE_ENV=development webpack -d --watch",
+    "compile": "babili src/main -d app --copy-files && cross-env NODE_ENV=production webpack -p",
     "package": "node ./packager.js",
     "lint": "eslint ./ui --ext .js --cache",
     "dist": "yarn compile && yarn package",
@@ -26,10 +27,12 @@
   "devDependencies": {
     "@blueprintjs/core": "^1.13.0",
     "algoliasearch": "^3.22.1",
+    "babel-cli": "^6.24.0",
     "babel-core": "^6.24.0",
     "babel-loader": "^6.3.2",
     "babel-preset-es2015": "^6.24.0",
     "babel-preset-react": "^6.16.0",
+    "babili": "^0.0.12",
     "chai": "^3.5.0",
     "chai-as-promised": "^6.0.0",
     "classnames": "^2.2.5",
