+++ src/main/libs/setProtocols.js
@@ -1,0 +1,7 @@
+const { app } = require('electron');
+
+const setProtocols = () => {
+  app.setAsDefaultProtocolClient('webcatalog');
+};
+
+module.exports = setProtocols;
